Improve the structure of the REST response for tasks.
The task description is now a embedded in the task resource as notes when requesting a single task.  Tasks in lists have a restricted set of fields.

Move the endpoint for creating a new task from /tasks to /projects/{id}/tasks.

Respond with CREATED rather than ACCEPTED when creating a new resource.